refactor: 🔄 Mutable.ts => types.ts
